// src/app/model/SchemeDetail.ts
export class SchemeDetail {
    detailId: number = 0;
    schemeImage: string = '';
    description: string = '';
    minAmount: number = 0;
    maxAmount: number = 0;
    minAge: number = 0;
    maxAge: number = 0;
    minTerm: number = 0;
    maxTerm: number = 0;
    profitPercent: number = 0;
    firstPremiumCommissionPercent: number = 0;
    emiCommissionPercent: number = 0;



    // status: boolean = true;
    // insuranceScheme: InsuranceScheme | null = null;
}